% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_summary.R
\name{annual_summary}
\alias{annual_summary}
\title{Annual Summary Statistics}
\usage{
annual_summary(daily_rfe, years, start_doy, end_doy, stat, outlocation)
}
\arguments{
\item{daily_rfe}{Folder containing the downloaded daily-rfe from tamsat_daily_download}

\item{years}{Years as interger(s) for which statistics will be calculated}

\item{start_doy}{Day of Year (DDD-format) for the start of the period}

\item{end_doy}{Day of Year (DDD-format) for the end of the period}

\item{stat}{statistic to calculate, must be compatible with raster::calc()}

\item{outlocation}{Folder (string) to save summary layers into}
}
\description{
Calculate statistical summaries of rainfall over a user-defined period
}

