% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_summary_from_months.R
\name{annual_summary_from_months}
\alias{annual_summary_from_months}
\title{Annual; Sum Statistics}
\usage{
annual_summary_from_months(download_folder, years, fill_with)
}
\arguments{
\item{download_folder}{Folder containing the downloaded monthly estimate from tamsat_download_all()}

\item{years}{Years as interger(s) for which statistics will be calculated}

\item{fill_with}{raster brick output from monthly_summary() to fill missing months}
}
\description{
Calculates sum annual rainfall from monthly estimate, will crop output to extent of fill_with
}

