% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_monthly_stack.R
\name{monthly_stack}
\alias{monthly_stack}
\title{Make complete one-month time step raster time series}
\usage{
monthly_stack(download_folder, years, fill_with)
}
\arguments{
\item{download_folder}{Folder containing the downloaded monthly estimate from tamsat_download_all()}

\item{years}{Years as interger(s) for which statistics will be calculated}

\item{fill_with}{raster brick output from monthly_summary() to fill missing months}
}
\description{
Read and stack raster layers, fill missing months with predefined averages
}

