% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_summary.R
\name{monthly_summary}
\alias{monthly_summary}
\title{Monthy; Summary Statistics}
\usage{
monthly_summary(download_folder, months, period, stat)
}
\arguments{
\item{download_folder}{Folder containing the downloaded monthly estimates from tamsat_all_download}

\item{months}{months to calculate eg c(1:12), c(1,4,6,10:12)}

\item{period}{Years as interger(s) for which statistics will be calculated}

\item{stat}{statistic to calculate, must be compatible with raster::calc()}
}
\description{
Calculates monthly statistical summaries of rainfall for each month for a defined period
}
\details{
Note: There seems to be some problem with some 1983 layers, for ease stick to 1984+
}

