% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_rainy_days.R, R/no_rainy_days_annual.R, R/no_rainy_days_window.R
\name{no_rainy_days_window}
\alias{no_rainy_days_window}
\title{Number of rainy days window}
\usage{
no_rainy_days_window(daily_rfe, years, start_doy, end_doy, threshold, id,
  outlocation)

no_rainy_days_window(daily_rfe, years, start_doy, end_doy, threshold, id,
  outlocation)

no_rainy_days_window(daily_rfe, years, start_doy, end_doy, threshold, id,
  outlocation)
}
\arguments{
\item{daily_rfe}{Folder containing the downloaded daily-rfe from tamsat_daily_download}

\item{years}{Years as interger(s) for which statistics will be calculated}

\item{start_doy}{Day of Year (DDD-format) for the start of the period}

\item{end_doy}{Day of Year (DDD-format) for the end of the period}

\item{threshold}{threshold for classifying a rainy day, in mm/day, defaults to zero is missing}

\item{id}{Charecter sting to identify the outputs eg "long_integer/ jan_march etc"}

\item{outlocation}{Folder (string) to save summary layers into}

\item{daily_rfe}{Folder containing the downloaded daily-rfe from tamsat_daily_download}

\item{threshold}{threshold for classifying a rainy day, in mm/day, defaults to zero is missing}

\item{years}{Years as interger(s) for which statistics will be calculated}

\item{id}{Charecter sting to identify the outputs eg "long_integer/ jan_march etc"}

\item{outlocation}{Folder (string) to save summary layers into}

\item{daily_rfe}{Folder containing the downloaded daily-rfe from tamsat_daily_download}

\item{start_doy}{Day of Year (DDD-format) for the start of the period}

\item{end_doy}{Day of Year (DDD-format) for the end of the period}

\item{threshold}{threshold for classifying a rainy day, in mm/day, defaults to zero is missing}

\item{years}{Years as interger(s) for which statistics will be calculated}

\item{id}{Charecter sting to identify the outputs eg "long_integer/ jan_march etc"}

\item{outlocation}{Folder (string) to save summary layers into}
}
\description{
Calculate number of days where rainfall (above a threshold) occurs

Calculate number of days where rainfall (above a threshold) occurs within a calender year

Calculate number of days where rainfall (above a threshold) occurs
}

