% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tamsat_daily_download.R
\name{tamsat_daily_download}
\alias{tamsat_daily_download}
\title{Download Tamsat Daily Rainfall Estimate}
\usage{
tamsat_daily_download(years, outlocation, unZip)
}
\arguments{
\item{years}{Years as interger(s) for which data will be downloaded}

\item{outlocation}{Folder (string) to save downloaded .zip files into}

\item{unZip}{Should the downloaded zips be unzipped}
}
\description{
Automate downloading of daily RFE's from the TAMSAT server
Note that the unZip utilities within R are particually slow
}

