% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_summary.R, R/annual_summary_from_daily.R
\name{annual_summary}
\alias{annual_summary}
\title{Annual; Summary Statistics}
\usage{
annual_summary(daily_rfe, years, stat, id, outlocation)

annual_summary(daily_rfe, years, stat, id, outlocation)
}
\arguments{
\item{daily_rfe}{Folder containing the downloaded daily-rfe from tamsat_daily_download}

\item{years}{Years as interger(s) for which statistics will be calculated}

\item{stat}{statistic to calculate, must be compatible with raster::calc()}

\item{id}{Charecter sting to identify the outputs eg 'long_integer/ jan_march etc'}

\item{outlocation}{Folder (string) to save summary layers into}

\item{daily_rfe}{Folder containing the downloaded daily-rfe from tamsat_daily_download}

\item{stat}{statistic to calculate, must be compatible with raster::calc()}

\item{years}{Years as interger(s) for which statistics will be calculated}

\item{id}{Charecter sting to identify the outputs eg 'long_integer/ jan_march etc'}

\item{outlocation}{Folder (string) to save summary layers into}
}
\description{
Calculate statistical summaries of rainfall over a calender year

Calculate statistical summaries of rainfall over a calender year
}

