% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tamsat_download_all.R
\name{tamsat_all_download}
\alias{tamsat_all_download}
\title{Download Tamsat Rainfall Estimate}
\usage{
tamsat_all_download(type, period, outlocation, unZip)
}
\arguments{
\item{type}{string, data to be downloaded of "e", (estimates) "a" (annomalies) or "both", defaults to both}

\item{period}{string, temporal resolution of data to download of "d" (daily), "m" (monthly), "s" (seasonal),or "all", defaults to all}

\item{outlocation}{Folder (string) to save downloaded .zip files into}

\item{unZip}{Should the downloaded zips be unzipped}
}
\description{
Automate downloading of daily RFE's from the TAMSAT server
Note that the unZip utilities within R are particually slow
}

